langchain/langchain_intro/requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
python-dotenv
openai
colorama